services/data-collector/requirements.txt
# MarketPrism Python Collector Dependencies

# 异步HTTP和WebSocket客户端
aiohttp==3.9.1
websockets==12.0

# NATS客户端
nats-py==2.6.0

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 日志记录
structlog==23.2.0
colorlog==6.8.0

# 数据处理
pandas==2.1.4
numpy==1.26.2

# 类型提示和验证
pydantic==2.5.1
typing-extensions==4.8.0

# 监控和指标
prometheus-client==0.19.0

# 工具库
click==8.1.7
uvloop==0.19.0  # Linux/macOS性能优化
orjson==3.9.10  # 高性能JSON处理

# 开发依赖
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 加密货币交易所API库
ccxt==4.1.64

# 时间处理
arrow==1.3.0 